requirements.txt
google-cloud-aiplatform
sentence-transformers
serpapi
python-dotenv
requests